+++ plugins/dot_viewer/src/dot_graphics_view.cpp
@@ -13,7 +13,7 @@ namespace hal {
 
     void DotGraphicsView::wheelEvent(QWheelEvent* event)
     {
-        qreal scaleFactor = qPow(2.0, event->delta() / 240.0); //How fast we zoom
+        qreal scaleFactor = qPow(2.0, event->angleDelta().y() / 240.0); //How fast we zoom
         scaleWithinLimits(scaleFactor);
 
     }
@@ -37,7 +37,7 @@ namespace hal {
     void DotGraphicsView::mousePressEvent(QMouseEvent* event)
     {
         if ((event->modifiers() == mPanModifier && event->button() == Qt::LeftButton) ||
-            (event->button() == Qt::MidButton && gGraphContextManager->sSettingPanOnMiddleButton->value().toBool()))
+            (event->button() == Qt::MiddleButton && gGraphContextManager->sSettingPanOnMiddleButton->value().toBool()))
         {
             mMovePosition = event->pos();
         }
@@ -49,7 +49,7 @@ namespace hal {
             return;
 
         if ((event->buttons().testFlag(Qt::LeftButton) && event->modifiers() == mPanModifier) ||
-            (event->buttons().testFlag(Qt::MidButton) && gGraphContextManager->sSettingPanOnMiddleButton->value().toBool()))
+            (event->buttons().testFlag(Qt::MiddleButton) && gGraphContextManager->sSettingPanOnMiddleButton->value().toBool()))
         {
             QScrollBar* hBar  = horizontalScrollBar();
             QScrollBar* vBar  = verticalScrollBar();