+++ plugins/gui/src/graph_widget/layouters/wait_to_be_seated.cpp
@@ -8,6 +8,7 @@
 #include <QDebug>
 #include <QMap>
 #include <QTextStream>
+#include <QIODevice>
 
 namespace hal
 {