+++ plugins/gui/src/python/python_code_editor.cpp
@@ -5,9 +5,9 @@
 #include "hal_core/netlist/project_manager.h"
 #include "hal_core/utilities/log.h"
 #include "gui/gui_globals.h"
+#include "gui/hal_qt_compat/hal_qt_compat.h"
 
 #include <QApplication>
-#include <QDesktopWidget>
 #include <QTextDocumentFragment>
 #include <QVBoxLayout>
 #include <QShortcut>
@@ -155,7 +155,7 @@ namespace hal
                 auto menu_width  = dialog->width();
                 auto menu_height = dialog->height();
 
-                auto desk_rect   = QApplication::desktop()->screenGeometry();
+                auto desk_rect   = QtCompat::desktopScreenGeometry();
                 auto desk_width  = desk_rect.width();
                 auto desk_height = desk_rect.height();
                 auto anchor      = this->cursorRect().bottomRight();